Move addEndpoint middleware
